+++ nuvla/job_engine/connector/helm_driver.py
@@ -7,7 +7,8 @@ from urllib.parse import urlparse
 
 from nuvla.job_engine.connector.connector import should_connect
 from nuvla.job_engine.connector.k8s_driver import Kubernetes
-from nuvla.job_engine.connector.utils import execute_cmd, create_tmp_file
+from nuvla.job_engine.connector.utils import (execute_cmd, create_tmp_file,
+                                              close_file)
 
 log = logging.getLogger('helm_driver')
 
@@ -16,6 +17,9 @@ class Helm:
     """
     Class to interact with Helm CLI.
     """
+
+    DOCKER_IO_HELM_AUTHN_URL = 'https://index.docker.io/v1/'
+
     def __init__(self, path_to_k8s_creds: str, **kwargs):
         self.k8s = Kubernetes.from_path_to_k8s_creds(path_to_k8s_creds, **kwargs)
         log.debug(self.k8s)
@@ -89,7 +93,8 @@ spec:
         return execute_cmd(cmd, env=env)
 
     def _from_absolute_url(self, op: str, release: str,
-                           absolute_url: str, namespace: str, values_yaml=None):
+                           absolute_url: str, namespace: str,
+                           values_yaml=None, work_dir=None) -> CompletedProcess:
         cmd = [op, release, absolute_url,
                '--namespace', namespace]
 
@@ -98,65 +103,57 @@ spec:
 
         values_yaml_fd = None
         if values_yaml:
-            values_yaml_fd = create_tmp_file(values_yaml)
+            values_yaml_fd = create_tmp_file(values_yaml, dir=work_dir)
             cmd += ['--values', values_yaml_fd.name]
         result = self.run_command(cmd)
         log.debug('Helm %s command result: %s', op, result)
-        if values_yaml_fd:
-            values_yaml_fd.close()
+        close_file(values_yaml_fd)
         return result
 
-    def _from_helm_repo_cred(self, op, helm_repo_url, helm_repo_cred: dict,
-                             chart_name, version, values_yaml, helm_release,
-                             namespace):
+    def _from_helm_repo_cred(self, op, helm_release, helm_repo_url,
+                             helm_repo_cred: dict, chart_name, version,
+                             namespace, values_yaml, work_dir=None):
         repos_config_fd = None
         registry_config_fd = None
-        env = {}
-
-        cmd = [op, helm_release, os.path.join(helm_repo_url, chart_name)]
 
         if helm_repo_url.startswith('http'):
+            repo_name = 'helm-repo'
+            cmd = [op, helm_release, f'{repo_name}/{chart_name}']
             cmd, repos_config_fd = self.__store_repo_config(
-                cmd, helm_repo_cred, helm_repo_url)
+                cmd, helm_repo_cred, helm_repo_url, repo_name, work_dir)
         elif helm_repo_url.startswith('oci'):
-            registry_config_fd = self.__store_registry_config(
-                env, helm_repo_cred, helm_repo_url)
+            cmd = [op, helm_release, os.path.join(helm_repo_url, chart_name)]
+            cmd, registry_config_fd = self.__store_registry_config(
+                cmd, helm_repo_cred, helm_repo_url, work_dir)
         else:
             raise Exception(f'Unsupported Helm repository URL: {helm_repo_url}')
 
-        cmd += ['--namespace', namespace]
-        if op == 'install' and namespace:
-            cmd += ['--create-namespace']
+        cmd += ['--namespace', namespace, '--create-namespace']
 
         if version:
             cmd += ['--version', version]
 
         values_yaml_fd = None
         if values_yaml:
-            values_yaml_fd = create_tmp_file(values_yaml)
+            values_yaml_fd = create_tmp_file(values_yaml, dir=work_dir)
             cmd += ['--values', values_yaml_fd.name]
 
         try:
-            result = self.run_command(cmd, env=env)
+            result = self.run_command(cmd)
             log.debug('Helm %s command result: %s', op, result)
         except Exception as e:
             log.error(f'Error running helm {op}: {e}')
             raise e
         else:
-            if repos_config_fd:
-                repos_config_fd.close()
-            if registry_config_fd:
-                registry_config_fd.close()
-            if values_yaml_fd:
-                values_yaml_fd.close()
+            for fd in [repos_config_fd, registry_config_fd, values_yaml_fd]:
+                close_file(fd)
 
         return result
 
     def __store_repo_config(self, cmd: list, helm_repo_cred: dict,
-                            helm_repo_url: str):
-        repo_name = 'helm-repo'
+                            helm_repo_url: str, repo_name: str, work_dir=None):
         repos_config = f"""
-apiVersion: ""
+apiVersion: v1
 generated: "0001-01-01T00:00:00Z"
 repositories:
 - caFile: ""
@@ -169,7 +166,7 @@ repositories:
   username: {helm_repo_cred['username']}
   password: {helm_repo_cred['password']}
 """
-        repos_config_fd = create_tmp_file(repos_config)
+        repos_config_fd = create_tmp_file(repos_config, dir=work_dir)
         result = self.run_command(['repo', 'update',
                                    '--repository-config', repos_config_fd.name])
         log.debug('Helm repo update command result: %s', result)
@@ -177,60 +174,64 @@ repositories:
         return cmd, repos_config_fd
 
     @staticmethod
-    def __store_registry_config(env: dict, helm_repo_cred: dict,
-                                helm_repo_url: str):
+    def __store_registry_config(cmd: list, helm_repo_cred: dict,
+                                helm_repo_url: str, work_dir=None):
         auth_string = base64.b64encode(
             f"{helm_repo_cred['username']}:{helm_repo_cred['password']}".encode(
                 'utf-8')
         ).decode('utf-8')
+        authn_url = urlparse(helm_repo_url).hostname
+        if authn_url.endswith('docker.io'):
+            authn_url = Helm.DOCKER_IO_HELM_AUTHN_URL
         registry_config = {
             "auths": {
-                urlparse(helm_repo_url).hostname: {
+                authn_url: {
                     "auth": auth_string
                 }
             }
         }
-        registry_config_fd = create_tmp_file(json.dumps(registry_config))
-        env.update({'HELM_REGISTRY_CONFIG': registry_config_fd.name})
-        return registry_config_fd
+        registry_config_fd = create_tmp_file(json.dumps(registry_config),
+                                             dir=work_dir)
+        cmd += ['--registry-config', registry_config_fd.name]
+        return cmd, registry_config_fd
 
-    def _from_helm_repo(self, chart_name, helm_release, helm_repo_url,
-                        namespace, op, values_yaml, version):
+    def _from_helm_repo(self, op, helm_release, helm_repo_url, chart_name,
+                        version, namespace, values_yaml, work_dir=None):
         if helm_repo_url.startswith('http'):
             cmd = [op, '--repo', helm_repo_url, helm_release, chart_name]
         elif helm_repo_url.startswith('oci'):
             cmd = [op, helm_release, os.path.join(helm_repo_url, chart_name)]
         else:
             raise Exception(f'Unsupported Helm repository URL: {helm_repo_url}')
-        cmd += ['--namespace', namespace]
-        if op == 'install' and namespace:
-            cmd += ['--create-namespace']
+
+        cmd += ['--namespace', namespace, '--create-namespace']
+
         if version:
             cmd += ['--version', version]
         values_yaml_fd = None
         if values_yaml:
-            values_yaml_fd = create_tmp_file(values_yaml)
+            values_yaml_fd = create_tmp_file(values_yaml, work_dir)
             cmd += ['--values', values_yaml_fd.name]
         result = self.run_command(cmd)
         log.debug('Helm %s command result: %s', op, result)
-        if values_yaml_fd:
-            values_yaml_fd.close()
+        close_file(values_yaml_fd)
         return result
 
     def op_install_upgrade(self, op, helm_release, helm_repo_url, helm_repo_cred,
                            helm_absolute_url, chart_name, version, namespace,
-                           values_yaml):
+                           values_yaml, work_dir=None):
         if helm_absolute_url:
             return self._from_absolute_url(op, helm_release, helm_absolute_url,
-                                           namespace)
+                                           namespace, values_yaml, work_dir)
         elif helm_repo_cred:
-            return self._from_helm_repo_cred(op, helm_repo_url, helm_repo_cred,
-                                             chart_name, version, values_yaml,
-                                             helm_release, namespace)
+            return self._from_helm_repo_cred(op, helm_release, helm_repo_url,
+                                             helm_repo_cred, chart_name,
+                                             version, namespace, values_yaml,
+                                             work_dir)
         else:
-            return self._from_helm_repo(chart_name, helm_release,
-                                        helm_repo_url, namespace, op,
-                                        values_yaml, version)
+            return self._from_helm_repo(op, helm_release, helm_repo_url,
+                                        chart_name, version, namespace,
+                                        values_yaml, work_dir)
 
     def install(self, helm_repo, helm_release, chart_name, namespace,
                 version=None, helm_repo_cred=dict, helm_absolute_url=None,
@@ -238,7 +239,7 @@ repositories:
         return self.op_install_upgrade('install', helm_release, helm_repo,
                                        helm_repo_cred, helm_absolute_url,
                                        chart_name, version, namespace,
-                                       chart_values_yaml)
+                                       chart_values_yaml, work_dir='/tmp')
 
     def upgrade(self, helm_repo, helm_release, chart_name, namespace,
                 version=None, helm_repo_cred=dict, helm_absolute_url=None,