+++ nuvla/job_engine/job/actions/bulk_update_deployment.py
@@ -4,12 +4,13 @@ import logging
 from ..actions import action
 from .utils.bulk_deployment import DeploymentBulkJob
 
+action_name = 'bulk_update_deployment'
 
-@action('bulk_update_deployment')
+@action(action_name)
 class DeploymentBulkUpdateJob(DeploymentBulkJob):
 
     def __init__(self, job):
-        super().__init__(job)
+        super().__init__(job, action_name)
 
     def deployment_action(self, deployment):
         module_href = self.job.payload.get('module-href')